# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI HR Agent Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/hr/jobs": {
      "get": {
        "tags": [
          "HR Jobs"
        ],
        "summary": "Get My Jobs",
        "operationId": "get_my_jobs_api_v1_hr_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobResponse"
                  },
                  "type": "array",
                  "title": "Response Get My Jobs Api V1 Hr Jobs Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "HR Jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_v1_hr_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/hr/jobs/{job_id}": {
      "get": {
        "tags": [
          "HR Jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_hr_jobs__job_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "HR Jobs"
        ],
        "summary": "Update Job",
        "operationId": "update_job_api_v1_hr_jobs__job_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "HR Jobs"
        ],
        "summary": "Delete Job",
        "operationId": "delete_job_api_v1_hr_jobs__job_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "Candidate Jobs"
        ],
        "summary": "Get Job Detail",
        "operationId": "get_job_detail_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/results/{application_id}": {
      "get": {
        "tags": [
          "Screening Results"
        ],
        "summary": "Get Application Result",
        "operationId": "get_application_result_api_v1_results__application_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/job/{job_id}": {
      "get": {
        "tags": [
          "Screening Results"
        ],
        "summary": "Get Job Results",
        "operationId": "get_job_results_api_v1_results_job__job_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScreeningResultResponse"
                  },
                  "title": "Response Get Job Results Api V1 Results Job  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "employment_type": {
            "type": "string",
            "title": "Employment Type"
          },
          "experience_required": {
            "type": "string",
            "title": "Experience Required"
          },
          "skills_required": {
            "type": "string",
            "title": "Skills Required"
          },
          "job_description": {
            "type": "string",
            "title": "Job Description"
          }
        },
        "type": "object",
        "required": [
          "title",
          "company_name",
          "location",
          "employment_type",
          "experience_required",
          "skills_required",
          "job_description"
        ],
        "title": "JobCreate"
      },
      "JobResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "employment_type": {
            "type": "string",
            "title": "Employment Type"
          },
          "experience_required": {
            "type": "string",
            "title": "Experience Required"
          },
          "skills_required": {
            "type": "string",
            "title": "Skills Required"
          },
          "job_description": {
            "type": "string",
            "title": "Job Description"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "hr_id": {
            "type": "integer",
            "title": "Hr Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "company_name",
          "location",
          "employment_type",
          "experience_required",
          "skills_required",
          "job_description",
          "id",
          "hr_id",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse"
      },
      "JobUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "employment_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employment Type"
          },
          "experience_required": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Required"
          },
          "skills_required": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills Required"
          },
          "job_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Description"
          }
        },
        "type": "object",
        "title": "JobUpdate"
      },
      "ScreeningResultResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "application_id": {
            "type": "integer",
            "title": "Application Id"
          },
          "semantic_score": {
            "type": "number",
            "title": "Semantic Score"
          },
          "skill_score": {
            "type": "number",
            "title": "Skill Score"
          },
          "experience_score": {
            "type": "number",
            "title": "Experience Score"
          },
          "final_score": {
            "type": "number",
            "title": "Final Score"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "strengths_json": {
            "type": "string",
            "title": "Strengths Json"
          },
          "missing_skills_json": {
            "type": "string",
            "title": "Missing Skills Json"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "explanation_text": {
            "type": "string",
            "title": "Explanation Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "application_id",
          "semantic_score",
          "skill_score",
          "experience_score",
          "final_score",
          "summary",
          "strengths_json",
          "missing_skills_json",
          "recommendation",
          "explanation_text",
          "created_at"
        ],
        "title": "ScreeningResultResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        }
      }
    }
  }
}